fraud_detection/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
watchdog==6.0.0
pandas==2.1.3
joblib==1.3.2 
scikit-learn==1.5.2
geoip2==5.0.1
